requirements.txt
alabaster==1.0.0
babel==2.17.0
beautifulsoup4==4.14.3
black==25.12.0
blinker==1.9.0
bs4==0.0.2
camelot-py==1.0.9
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.1
cryptography==46.0.3
docutils==0.22.4
et_xmlfile==2.0.0
fatsecret==0.5.0
Flask==3.1.2
idna==3.11
imagesize==1.4.1
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
pdfminer.six==20260107
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
pypdf==5.9.0
pypdfium2==5.3.0
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
rauth==0.7.3
requests==2.32.5
roman-numerals==4.1.0
ruff==0.14.13
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.8.1
Sphinx==9.1.0
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
